requirements.txt
fastapi[all]
langchain
langchain_groq
python-dotenv